Refactoring
Function '_patch_obj_resource_item' was introduced and placed into the
'APIBaseClient' class. Method 'upload_document' was moved from
'TendersClient' and 'ContractingClient' classes to parent class.
