+++ openprocurement_client/contract.py
@@ -15,17 +15,6 @@ class ContractingClient(APIBaseClient):
         super(ContractingClient, self).__init__(key, 'contracts', host_url,
                                                 api_version, params, ds_client)
 
-    @verify_file
-    def upload_document(self, file_, contract):
-        return self._upload_resource_file(
-            '{}/{}/documents'.format(
-                self.prefix_path,
-                contract.data.id
-            ),
-            files=file_,
-            headers={'X-Access-Token': self._get_access_token(contract)}
-        )
-
     def create_contract(self, contract):
         return self._create_resource_item(self.prefix_path, contract)
